requirements.txt
# Core packages
numpy>=1.20.0
torch>=1.9.0
transformers>=4.18.0
datasets>=2.0.0
pydantic>=1.9.0
tqdm>=4.62.0
pretty-traceback>=0.1.2
yaml-config>=0.1.7
matplotlib>=3.5.0

# Optional packages for extended functionality
scipy>=1.7.0
pandas>=1.3.0
wandb>=0.12.0  # For experiment tracking (optional)
